+++ CommunicationXLib/Common/source/CommunicationModel/CommonLayer/Classes/Commands/VirtualCommandSet/Command_VCS_Drive2.cpp
@@ -1685,9 +1685,9 @@ void CCommand_VCS_Drive2::SetDefaultParameter_GetHomingParameter()
     DWORD ulHomingAcceleration = 0;
     DWORD ulSpeedSwitch = 0;
     DWORD ulSpeedIndex = 0;
-    long lHomeOffset = 0;
+    int lHomeOffset = 0;
     WORD usCurrentThreshold = 0;
-    long lHomePosition = 0;
+    int lHomePosition = 0;
 
     //ReturnParameter
     SetReturnParameterData(0, &ulHomingAcceleration, sizeof(ulHomingAcceleration));
@@ -1703,9 +1703,9 @@ void CCommand_VCS_Drive2::SetDefaultParameter_SetHomingParameter()
     DWORD ulHomingAcceleration = 0;
     DWORD ulSpeedSwitch = 0;
     DWORD ulSpeedIndex = 0;
-    long lHomeOffset = 0;
+    int lHomeOffset = 0;
     WORD usCurrentThreshold = 0;
-    long lHomePosition = 0;
+    int lHomePosition = 0;
 
     //Parameter
     SetParameterData(0, &ulHomingAcceleration, sizeof(ulHomingAcceleration));
@@ -1723,7 +1723,7 @@ void CCommand_VCS_Drive2::SetDefaultParameter_StopHoming()
 
 void CCommand_VCS_Drive2::SetDefaultParameter_DefinePosition()
 {
-    long lHomePosition = 0;
+    int lHomePosition = 0;
 
     //Parameter
     SetParameterData(0, &lHomePosition, sizeof(lHomePosition));
@@ -1899,7 +1899,7 @@ void CCommand_VCS_Drive2::SetDefaultParameter_DisablePositionCompare()
 
 void CCommand_VCS_Drive2::SetDefaultParameter_SetPositionCompareReferencePosition()
 {
-    long lReferencePosition = 0;
+    int lReferencePosition = 0;
 
     //Parameter
     SetParameterData(0, &lReferencePosition, sizeof(lReferencePosition));
@@ -1958,7 +1958,7 @@ void CCommand_VCS_Drive2::SetDefaultParameter_ReadPositionMarkerCounter()
 void CCommand_VCS_Drive2::SetDefaultParameter_ReadPositionMarkerCapturedPosition()
 {
     WORD usCounterIndex = 0;
-    long lCapturedPosition = 0;
+    int lCapturedPosition = 0;
 
     //Parameter
     SetParameterData(0, &usCounterIndex, sizeof(usCounterIndex));
@@ -2004,7 +2004,7 @@ void CCommand_VCS_Drive2::SetDefaultParameter_GetMovementState()
 
 void CCommand_VCS_Drive2::SetDefaultParameter_GetPositionIs()
 {
-    long lPositionIs = 0;
+    int lPositionIs = 0;
 
     //Parameter
 
@@ -2014,7 +2014,7 @@ void CCommand_VCS_Drive2::SetDefaultParameter_GetPositionIs()
 
 void CCommand_VCS_Drive2::SetDefaultParameter_GetVelocityIs()
 {
-    long lVelocityIs = 0;
+    int lVelocityIs = 0;
 
     //Parameter
 
@@ -2024,7 +2024,7 @@ void CCommand_VCS_Drive2::SetDefaultParameter_GetVelocityIs()
 
 void CCommand_VCS_Drive2::SetDefaultParameter_GetVelocityIsAveraged()
 {
-    long lVelocityIsAveraged = 0;
+    int lVelocityIsAveraged = 0;
 
     //Parameter
 
@@ -2034,7 +2034,7 @@ void CCommand_VCS_Drive2::SetDefaultParameter_GetVelocityIsAveraged()
 
 void CCommand_VCS_Drive2::SetDefaultParameter_GetPositionMust()
 {
-    long lPositionMust = 0;
+    int lPositionMust = 0;
 
     //Parameter
 
@@ -2044,7 +2044,7 @@ void CCommand_VCS_Drive2::SetDefaultParameter_GetPositionMust()
 
 void CCommand_VCS_Drive2::SetDefaultParameter_SetPositionMust()
 {
-    long lPositionMust = 0;
+    int lPositionMust = 0;
 
     //Parameter
     SetParameterData(0, &lPositionMust, sizeof(lPositionMust));
@@ -2054,7 +2054,7 @@ void CCommand_VCS_Drive2::SetDefaultParameter_ActivateAnalogPositionSetpoint()
 {
     WORD usAnalogInputNumber = 0;
     float fScaling = 0;
-    long lOffset = 0;
+    int lOffset = 0;
 
     //Parameter
     SetParameterData(0, &usAnalogInputNumber, sizeof(usAnalogInputNumber));
@@ -2096,7 +2096,7 @@ void CCommand_VCS_Drive2::SetDefaultParameter_GetPositionProfile()
 
 void CCommand_VCS_Drive2::SetDefaultParameter_GetTargetPosition()
 {
-    long lTargetPosition = 0;
+    int lTargetPosition = 0;
 
     //Parameter
 
@@ -2111,7 +2111,7 @@ void CCommand_VCS_Drive2::SetDefaultParameter_HaltPositionMovement()
 
 void CCommand_VCS_Drive2::SetDefaultParameter_MoveToPosition()
 {
-    long lTargetPosition = 1000;
+    int lTargetPosition = 1000;
     BOOL oAbsolute = FALSE;
     BOOL oImmediately = TRUE;
 
@@ -2150,7 +2150,7 @@ void CCommand_VCS_Drive2::SetDefaultParameter_DisablePositionWindow()
 
 void CCommand_VCS_Drive2::SetDefaultParameter_GetTargetVelocity()
 {
-    long lTargetVelocity = 0;
+    int lTargetVelocity = 0;
 
     //Parameter
 
@@ -2177,7 +2177,7 @@ void CCommand_VCS_Drive2::SetDefaultParameter_HaltVelocityMovement()
 
 void CCommand_VCS_Drive2::SetDefaultParameter_MoveWithVelocity()
 {
-    long lTargetVelocity = 0;
+    int lTargetVelocity = 0;
 
     //Parameter
     SetParameterData(0, &lTargetVelocity, sizeof(lTargetVelocity));
@@ -2306,7 +2306,7 @@ void CCommand_VCS_Drive2::SetDefaultParameter_SetQuickStopState()
 
 void CCommand_VCS_Drive2::SetDefaultParameter_GetVelocityMust()
 {
-    long lVelocityMust = 0;
+    int lVelocityMust = 0;
 
     //Parameter
 
@@ -2316,7 +2316,7 @@ void CCommand_VCS_Drive2::SetDefaultParameter_GetVelocityMust()
 
 void CCommand_VCS_Drive2::SetDefaultParameter_SetVelocityMust()
 {
-    long lVelocityMust = 0;
+    int lVelocityMust = 0;
 
     //Parameter
     SetParameterData(0, &lVelocityMust, sizeof(lVelocityMust));
@@ -2326,7 +2326,7 @@ void CCommand_VCS_Drive2::SetDefaultParameter_ActivateAnalogVelocitySetpoint()
 {
     WORD usAnalogInputNumber = 0;
     float fScaling = 0;
-    long lOffset = 0;
+    int lOffset = 0;
 
     //Parameter
     SetParameterData(0, &usAnalogInputNumber, sizeof(usAnalogInputNumber));
@@ -2413,8 +2413,8 @@ void CCommand_VCS_Drive2::SetDefaultParameter_GetFreeIpmBufferSize()
 
 void CCommand_VCS_Drive2::SetDefaultParameter_AddPvtValueToIpmBuffer()
 {
-    long lPosition = 0;
-    long lVelocity = 0;
+    int lPosition = 0;
+    int lVelocity = 0;
     BYTE ubTime = 0;
 
     //Parameter
